refactor: relocate ResolvedAgentTaskProviderPolicy to leaf schedule module
ResolvedAgentTaskProviderPolicy is a plain data struct, but it lived in
agent_task_dispatch_service alongside the dispatch machinery (which pulls in six
agent-task subsystem modules). The lab-contract type layer
(command_contract::lab::workload) held it as a field, dragging that entire
service dependency onto the shared contract surface and feeding the
core <-> command_contract cycle.

Move the struct into the leaf agent_task_schedule::config module beside its
AgentTaskProviderRotationPolicy / AgentTaskRetryPolicy fields (zero core deps),
re-export it from agent_task_dispatch_service to keep existing call sites stable,
and point lab/workload.rs at the leaf location. workload.rs no longer depends on
agent_task_dispatch_service.

Second cut toward breaking the core <-> command_contract cycle.
